Split memoize into own util
