+++ src/utils/memoize.ts
@@ -1,0 +1,12 @@
+export function memoize<T extends (...args: any[]) => any>(fn: T): T {
+	const cache = new Map<string, ReturnType<T>>();
+	return ((...args: Parameters<T>): ReturnType<T> => {
+		const key = JSON.stringify(args);
+		if (cache.has(key)) {
+			return cache.get(key)!;
+		}
+		const result = fn(...args);
+		cache.set(key, result);
+		return result;
+	}) as T;
+}
